requirements.txt
Django>=4.0.4,<4.1
psycopg2>=2.9.3,<2.10
djangorestframework>=3.13.1,<3.14
drf-spectacular>=0.22.1,<0.23

